pyproject.toml
[tool.poetry]
name = "devchat"
version = "0.1.0"
description = "CodeAssist is an AI-powered coding assistant that helps developers write better code through intelligent workflows and chat-based interactions."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/devchat"
packages = [
    { include = "devchat", from = "src" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Debuggers"
]

[tool.poetry.dependencies]
python = "^3.8.1"
pydantic = "1.10.16"
networkx = "^3.1"
openai = "1.35.15"
rich_click = "^1.6.1"
tiktoken = ">0.4.0"
tinydb = "^4.7.1"
urllib3 = "<2.0"
importlib-metadata = "^6.8.0"
gitpython = "^3.1.32"
oyaml = "^1.0"
colorama = "^0.4.6"
tenacity = "^8.2.3"
pathspec = "^0.12.1"
importlib-resources = "^6.1.1"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
gunicorn = "^22.0.0"
loguru = "^0.7.2"
win32-setctime = "^1.1.0"
virtualenv = "^20.27.1"
# LangChain dependencies
langchain = "^0.1.12"
langchain-core = "^0.1.31"
langchain-openai = "^0.0.8"
# Additional utilities
python-dotenv = "^1.0.1"
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"
pylint = "^3.0.3"
astroid = "^3.0.2"

[tool.poetry.scripts]
devchat = "devchat.cli.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
ruff = "^0.4.4"
black = "^23.0.0"
isort = "^5.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # Error
    "W",  # Warning
    "F",  # pyflakes
    "I",  # isort
]
fixable = ["ALL"]

